requirements.txt
fastapi
uvicorn
python-chess
jinja2
websockets
google-generativeai
python-dotenv